AIUtiils/requirements.txt
requests==2.32.3
pandas==2.2.3
numpy==2.2.2
pymongo==4.10.1
scikit-learn==1.6.1
PyYAML==6.0.2
pytest==8.3.4
scipy==1.15.1
imblearn==0.0
dill==0.3.4